requirements.txt
asgiref==3.8.1              # For async support with Django
django==5.1.3               # Django framework for the web application
django-cors-headers==4.6.0  # To handle Cross-Origin Resource Sharing (CORS)
python-dotenv==1.0.1        # For loading environment variables from .env file
PyPDF2==3.0.1               # To extract text from PDF files
openai==1.58.1              # OpenAI API client
pinecone-client==5.0.1      # Pinecone API client
langchain==0.3.13           # LangChain for LLM workflows
langchain-community==0.3.13 # Additional community-driven LangChain integrations
langchain-core==0.3.28      # Core LangChain utilities
langchain-openai==0.2.14    # LangChain OpenAI support
langchain-pinecone==0.2.0   # LangChain Pinecone integration
uvicorn==0.32.1             # ASGI server to run the application
langgraph>=0.0.1            # StateGraph utility used for workflow definition
boto3==1.28.50              # AWS SDK for Python for S3 and other AWS services
Pillow==9.5.0               # For image validation and conversion